refactor(dashboard): move SplitView into (auth)/
