+++ ui/apps/dashboard/src/app/(auth)/SplitView.tsx
